Added `PropertyListener`, a `ChangeListener<Object>`, to make listening to widget properties easier.
