+++ lib/src/main/java/com/wjduquette/joe/win/Listener.java
@@ -1,0 +1,11 @@
+package com.wjduquette.joe.win;
+
+/**
+ * A cancelable listener.
+ */
+public interface Listener {
+    /**
+     * Cancels the listener, removing it from the list of listeners.
+     */
+    void cancel();
+}
